requirements.txt
Flask==3.1.1
flask-cors==6.0.1
pymongo==4.13.2
python-dotenv==1.1.1
Flask-PyMongo==3.0.1
